:truck: Rename and move the suggested other resources section
I like the idea. I had always thought of this document as being one about
libraries and applications, but I can see this being useful too. For now
though, given there's just the one entry, and given tutorials can
potentially be more ephemeral, I think it's better to not emphasise this
section quite so much.

Perhaps when Textual settles down and the collection of such links grows
I'll bump it up the document.